[BUGFIX] #5505 Drop autoload-dev
The autoload-dev section in the composer.json has no effect. However, the
autoload-dev section from the ext_emconf.php (which works fine) is dropped
by the TER.

So we need to move the autoloading of the test classes from autoload-dev
to autoload in the composer.json so that developers can run the tests
even with the TER version of this extension.

Change-Id: Ia1b4d3a7eb1c1a4f83ea465c5abcaf704dca7ea6
Reviewed-on: https://review.typo3.org/45981
Reviewed-by: Stefano Kowalke <blueduck@mailbox.org>
Tested-by: Stefano Kowalke <blueduck@mailbox.org>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>